+++ engine/shared/library/sharedFoundationTypes/src/linux/FoundationTypesLinux.h
@@ -16,10 +16,10 @@
 typedef unsigned int           uint;
 typedef unsigned char          uint8;
 typedef unsigned short         uint16;
-typedef unsigned long          uint32;
+typedef unsigned int           uint32;
 typedef signed char            int8;
 typedef signed short           int16;
-typedef signed long            int32;
+typedef signed int             int32;
 typedef signed long long int   int64;
 typedef unsigned long long int uint64;
 typedef float                  real;
